requirements.txt
langchain>=0.0.190
langchain-google-genai>=0.1.0
langchain-community>=0.1.20
pypdf>=3.14.0
faiss-cpu>=1.7.3
streamlit>=1.24.1
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
